+++ luarocks/src/luarocks/fetch/svn.lua
@@ -1,6 +1,9 @@
 
 --- Fetch back-end for retrieving sources from Subversion.
-module("luarocks.fetch.svn", package.seeall)
+--module("luarocks.fetch.svn", package.seeall)
+local svn = {}
+
+local unpack = unpack or table.unpack
 
 local fs = require("luarocks.fs")
 local dir = require("luarocks.dir")
@@ -13,7 +16,7 @@ local util = require("luarocks.util")
 -- @return (string, string) or (nil, string): The absolute pathname of
 -- the fetched source tarball and the temporary directory created to
 -- store it; or nil and an error message.
-function get_sources(rockspec, extract, dest_dir)
+function svn.get_sources(rockspec, extract, dest_dir)
    assert(type(rockspec) == "table")
    assert(type(dest_dir) == "string" or not dest_dir)
 
@@ -36,11 +39,13 @@ function get_sources(rockspec, extract, dest_dir)
    else
       store_dir = dest_dir
    end
-   fs.change_dir(store_dir)
+   local ok, err = fs.change_dir(store_dir)
+   if not ok then return nil, err end
    if not fs.execute(unpack(command)) then
       return nil, "Failed fetching files from Subversion."
    end
-   fs.change_dir(module)
+   ok, err = fs.change_dir(module)
+   if not ok then return nil, err end
    for _, d in ipairs(fs.find(".")) do
       if dir.base_name(d) == ".svn" then
          fs.delete(dir.path(store_dir, module, d))
@@ -51,3 +56,5 @@ function get_sources(rockspec, extract, dest_dir)
    return module, store_dir
 end
 
+
+return svn
